+++ CHANGES.md
@@ -17,6 +17,7 @@
 - use the UniProt accession number if the GN field is not specific enough
 - when UniProt records point to NCBI gene records, obtain the nucleotide accession from the gene record, and protein ID from the RefSeq cross-reference
 - use the UniProt ORF gene name field for identifying CDS in larger records, to avoid issues with ambiguous GN fields
+- adapt to bioservices 1.12 change that modifies UniProt return string format
 
 ## v0.2.0